+++ client/Assets/Scripts/BasicSystem/NBT/Blocks/NBTSnowLayer.cs
@@ -87,7 +87,7 @@ public class NBTSnowLayer : NBTBlock
 
     protected override FaceAttributes GetFrontFaceAttributes(NBTChunk chunk, NBTMesh mesh, CubeAttributes ca)
     {
-        chunk.GetLights(ca.pos.x, ca.pos.y, ca.pos.z - 1, out byte skyLight, out byte blockLight);
+        chunk.GetLights(ca.pos.x, ca.pos.y, ca.pos.z - 1, out float skyLight, out float blockLight);
 
         FaceAttributes fa = new FaceAttributes();
         fa.pos = frontVertices_snow;
@@ -102,7 +102,7 @@ public class NBTSnowLayer : NBTBlock
     }
     protected override FaceAttributes GetBackFaceAttributes(NBTChunk chunk, NBTMesh mesh, CubeAttributes ca)
     {
-        chunk.GetLights(ca.pos.x, ca.pos.y, ca.pos.z + 1, out byte skyLight, out byte blockLight);
+        chunk.GetLights(ca.pos.x, ca.pos.y, ca.pos.z + 1, out float skyLight, out float blockLight);
 
         FaceAttributes fa = new FaceAttributes();
         fa.pos = backVertices_snow;
@@ -117,7 +117,7 @@ public class NBTSnowLayer : NBTBlock
     }
     protected override FaceAttributes GetTopFaceAttributes(NBTChunk chunk, NBTMesh mesh, CubeAttributes ca)
     {
-        chunk.GetLights(ca.pos.x, ca.pos.y + 1, ca.pos.z, out byte skyLight, out byte blockLight);
+        chunk.GetLights(ca.pos.x, ca.pos.y + 1, ca.pos.z, out float skyLight, out float blockLight);
 
         FaceAttributes fa = new FaceAttributes();
         fa.pos = topVertices_snow;
@@ -132,7 +132,7 @@ public class NBTSnowLayer : NBTBlock
     }
     protected override FaceAttributes GetBottomFaceAttributes(NBTChunk chunk, NBTMesh mesh, CubeAttributes ca)
     {
-        chunk.GetLights(ca.pos.x, ca.pos.y - 1, ca.pos.z, out byte skyLight, out byte blockLight);
+        chunk.GetLights(ca.pos.x, ca.pos.y - 1, ca.pos.z, out float skyLight, out float blockLight);
 
         FaceAttributes fa;
         fa.pos = bottomVertices_snow;
@@ -147,7 +147,7 @@ public class NBTSnowLayer : NBTBlock
     }
     protected override FaceAttributes GetLeftFaceAttributes(NBTChunk chunk, NBTMesh mesh, CubeAttributes ca)
     {
-        chunk.GetLights(ca.pos.x - 1, ca.pos.y, ca.pos.z, out byte skyLight, out byte blockLight);
+        chunk.GetLights(ca.pos.x - 1, ca.pos.y, ca.pos.z, out float skyLight, out float blockLight);
 
         FaceAttributes fa;
         fa.pos = leftVertices_snow;
@@ -162,7 +162,7 @@ public class NBTSnowLayer : NBTBlock
     }
     protected override FaceAttributes GetRightFaceAttributes(NBTChunk chunk, NBTMesh mesh, CubeAttributes ca)
     {
-        chunk.GetLights(ca.pos.x + 1, ca.pos.y, ca.pos.z, out byte skyLight, out byte blockLight);
+        chunk.GetLights(ca.pos.x + 1, ca.pos.y, ca.pos.z, out float skyLight, out float blockLight);
 
         FaceAttributes fa;
         fa.pos = rightVertices_snow;
