+++ client/Assets/Scripts/BasicSystem/NBT/Blocks/Stairs/NBTStairs.cs
@@ -32,7 +32,7 @@ public class NBTStairs : NBTBlock
 
         int faceIndex = TextureArrayManager.GetIndexByName(stairsName);
 
-        chunk.GetLights(pos.x, pos.y, pos.z, out byte skyLight, out byte blockLight);
+        chunk.GetLights(pos.x, pos.y, pos.z, out float skyLight, out float blockLight);
 
         NBTMesh nbtMesh = nbtGO.nbtMesh;
         ushort startIndex = nbtMesh.vertexCount;
